feat: commission report functionality removed
