--- src/main/java/satyamconsignment/ui/Main/WelcomeScreen/WelcomeScreenController.java
@@ -1,24 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package satyamconsignment.ui.Main.WelcomeScreen;
-
-import java.net.URL;
-import java.util.ResourceBundle;
-import javafx.fxml.Initializable;
-
-/**
- * FXML Controller class
- *
- * @author ADMIN
- */
-public class WelcomeScreenController implements Initializable {
-
-    /** Initializes the controller class. */
-    @Override
-    public void initialize(URL url, ResourceBundle rb) {
-        // TODO
-    }
-}
