+++ src/main/resources/satyamconsignment/ui/Main/CollectionHistory/CollectionHistory.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.text.Text?>
 <?import java.net.URL?>
 <Group xmlns:fx="http://javafx.com/fxml/1" fx:id="root2" xmlns="http://javafx.com/javafx/8.0.111"
-       fx:controller="satyamconsignment.ui.Main.CollectionHistory.CollectionHistoryController">
+       fx:controller="satyamconsignment.ui.Main.CollectionHistory.CollectionHistory">
     <children>
         <AnchorPane id="AnchorPane" prefHeight="511.0" prefWidth="600.0" styleClass="mainFxmlClass">
             <stylesheets>
