+++ src/main/resources/satyamconsignment/ui/Main/InputHistory/InputHistory.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.text.Text?>
 <?import java.net.URL?>
 <Group xmlns:fx="http://javafx.com/fxml/1" fx:id="root2" xmlns="http://javafx.com/javafx/8.0.111"
-       fx:controller="satyamconsignment.ui.Main.InputHistory.InputHistoryController">
+       fx:controller="satyamconsignment.ui.Main.InputHistory.InputHistory">
     <children>
         <AnchorPane id="AnchorPane" prefHeight="548.0" prefWidth="621.0" styleClass="mainFxmlClass">
             <stylesheets>
